Print the Elements of a Linked List
